// src/generated/stoporders.ts
export enum StopOrderDirection {
  /** STOP_ORDER_DIRECTION_UNSPECIFIED - Значение не указано */
  STOP_ORDER_DIRECTION_UNSPECIFIED = 0,
  /** STOP_ORDER_DIRECTION_BUY - Покупка */
  STOP_ORDER_DIRECTION_BUY = 1,
  /** STOP_ORDER_DIRECTION_SELL - Продажа */
  STOP_ORDER_DIRECTION_SELL = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum StopOrderType {
  /** STOP_ORDER_TYPE_UNSPECIFIED - Значение не указано */
  STOP_ORDER_TYPE_UNSPECIFIED = 0,
  /** STOP_ORDER_TYPE_TAKE_PROFIT - Take-profit заявка */
  STOP_ORDER_TYPE_TAKE_PROFIT = 1,
  /** STOP_ORDER_TYPE_STOP_LOSS - Stop-loss заявка */
  STOP_ORDER_TYPE_STOP_LOSS = 2,
  /** STOP_ORDER_TYPE_STOP_LIMIT - Stop-limit заявка */
  STOP_ORDER_TYPE_STOP_LIMIT = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum StopOrderStatusOption {
  /** STOP_ORDER_STATUS_UNSPECIFIED - Значение не указано */
  STOP_ORDER_STATUS_UNSPECIFIED = 0,
  /** STOP_ORDER_STATUS_ALL - Все заявки */
  STOP_ORDER_STATUS_ALL = 1,
  /** STOP_ORDER_STATUS_ACTIVE - Активные заявки */
  STOP_ORDER_STATUS_ACTIVE = 2,
  /** STOP_ORDER_STATUS_EXECUTED - Исполненные заявки */
  STOP_ORDER_STATUS_EXECUTED = 3,
  /** STOP_ORDER_STATUS_CANCELED - Отмененные заявки */
  STOP_ORDER_STATUS_CANCELED = 4,
  /** STOP_ORDER_STATUS_EXPIRED - Истекшие заявки */
  STOP_ORDER_STATUS_EXPIRED = 5,
  UNRECOGNIZED = -1,
}
// ...
export enum ExchangeOrderType {
  /** EXCHANGE_ORDER_TYPE_UNSPECIFIED - Значение не указано */
  EXCHANGE_ORDER_TYPE_UNSPECIFIED = 0,
  /** EXCHANGE_ORDER_TYPE_MARKET - Заявка по рыночной цене */
  EXCHANGE_ORDER_TYPE_MARKET = 1,
  /** EXCHANGE_ORDER_TYPE_LIMIT - Лимитная заявка */
  EXCHANGE_ORDER_TYPE_LIMIT = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TakeProfitType {
  /** TAKE_PROFIT_TYPE_UNSPECIFIED - Значение не указано */
  TAKE_PROFIT_TYPE_UNSPECIFIED = 0,
  /** TAKE_PROFIT_TYPE_REGULAR - Обычная заявка (значение по умолчанию) */
  TAKE_PROFIT_TYPE_REGULAR = 1,
  /** TAKE_PROFIT_TYPE_TRAILING - Трейлинг-стоп */
  TAKE_PROFIT_TYPE_TRAILING = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TrailingValueType {
  /** TRAILING_VALUE_UNSPECIFIED - Значение не указано */
  TRAILING_VALUE_UNSPECIFIED = 0,
  /** TRAILING_VALUE_ABSOLUTE - Абсолютное значение в единицах цены */
  TRAILING_VALUE_ABSOLUTE = 1,
  /** TRAILING_VALUE_RELATIVE - Относительное значение в процентах */
  TRAILING_VALUE_RELATIVE = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum TrailingStopStatus {
  /** TRAILING_STOP_UNSPECIFIED - Значение не указано */
  TRAILING_STOP_UNSPECIFIED = 0,
  /** TRAILING_STOP_ACTIVE - Активный */
  TRAILING_STOP_ACTIVE = 1,
  /** TRAILING_STOP_ACTIVATED - Активированный */
  TRAILING_STOP_ACTIVATED = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface GetStopOrdersRequest {
  /** Идентификатор счёта клиента */
  accountId: string;
  /** Статус заявок */
  status: StopOrderStatusOption;
  /** Левая граница */
  from:
    | Date
    | undefined;
  /** Правая граница */
  to: Date | undefined;
}

/** Список активных стоп-заявок. */
export interface GetStopOrdersResponse {
  /** Массив стоп-заявок по счёту */
  stopOrders: StopOrder[];
}
// ...
export interface StopOrder {
  /** Идентификатор-идентификатор стоп-заявки */
  stopOrderId: string;
  /** Запрошено лотов */
  lotsRequested: number;
  /** Figi-идентификатор инструмента */
  figi: string;
  /** Направление операции */
  direction: StopOrderDirection;
  /** Валюта стоп-заявки */
  currency: string;
  /** Тип стоп-заявки */
  orderType: StopOrderType;
  /** Дата и время выставления заявки в часовом поясе UTC */
  createDate:
    | Date
    | undefined;
  /** Дата и время конвертации стоп-заявки в биржевую в часовом поясе UTC */
  activationDateTime:
    | Date
    | undefined;
  /** Дата и время снятия заявки в часовом поясе UTC */
  expirationTime:
    | Date
    | undefined;
  /** Цена заявки за 1 инструмент. Для получения стоимости лота требуется умножить на лотность инструмента. */
  price:
    | MoneyValue
    | undefined;
  /** Цена активации стоп-заявки за 1 инструмент. Для получения стоимости лота требуется умножить на лотность инструмента. */
  stopPrice:
    | MoneyValue
    | undefined;
  /** instrument_uid идентификатор инструмента */
  instrumentUid: string;
  /** Подтип стоп-заявки TakeProfit */
  takeProfitType: TakeProfitType;
  /** Параметры трейлинг-стопа */
  trailingData:
    | StopOrder_TrailingData
    | undefined;
  /** Статус заявки */
  status: StopOrderStatusOption;
  /** Тип дочерней биржевой заявки для тейкпрофита */
  exchangeOrderType: ExchangeOrderType;
}

export interface StopOrder_TrailingData {
  /** Отступ */
  indent:
    | Quotation
    | undefined;
  /** Тип величины отступа */
  indentType: TrailingValueType;
  /** Размер защитного спреда */
  spread:
    | Quotation
    | undefined;
  /** Тип величины защитного спреда */
  spreadType: TrailingValueType;
  /** Статус трейлинг-стопа */
  status: TrailingStopStatus;
  /** Цена исполнения */
  price:
    | Quotation
    | undefined;
  /** Локальный экстремум */
  extr: Quotation | undefined;
}
// ...
export const GetStopOrdersRequest = {
  encode(message: GetStopOrdersRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.accountId !== "") {
      writer.uint32(10).string(message.accountId);
    }
    if (message.status !== 0) {
      writer.uint32(16).int32(message.status);
    }
    if (message.from !== undefined) {
      Timestamp.encode(toTimestamp(message.from), writer.uint32(26).fork()).ldelim();
    }
    if (message.to !== undefined) {
      Timestamp.encode(toTimestamp(message.to), writer.uint32(34).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetStopOrdersRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetStopOrdersRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.accountId = reader.string();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.from = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.to = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetStopOrdersRequest {
    return {
      accountId: isSet(object.accountId) ? globalThis.String(object.accountId) : "",
      status: isSet(object.status) ? stopOrderStatusOptionFromJSON(object.status) : 0,
      from: isSet(object.from) ? fromJsonTimestamp(object.from) : undefined,
      to: isSet(object.to) ? fromJsonTimestamp(object.to) : undefined,
    };
  },

  toJSON(message: GetStopOrdersRequest): unknown {
    const obj: any = {};
    if (message.accountId !== "") {
      obj.accountId = message.accountId;
    }
    if (message.status !== 0) {
      obj.status = stopOrderStatusOptionToJSON(message.status);
    }
    if (message.from !== undefined) {
      obj.from = message.from.toISOString();
    }
    if (message.to !== undefined) {
      obj.to = message.to.toISOString();
    }
    return obj;
  },

  create(base?: DeepPartial<GetStopOrdersRequest>): GetStopOrdersRequest {
    return GetStopOrdersRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetStopOrdersRequest>): GetStopOrdersRequest {
    const message = createBaseGetStopOrdersRequest();
    message.accountId = object.accountId ?? "";
    message.status = object.status ?? 0;
    message.from = object.from ?? undefined;
    message.to = object.to ?? undefined;
    return message;
  },
};

function createBaseGetStopOrdersResponse(): GetStopOrdersResponse {
  return { stopOrders: [] };
}

export const GetStopOrdersResponse = {
  encode(message: GetStopOrdersResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    for (const v of message.stopOrders) {
      StopOrder.encode(v!, writer.uint32(10).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetStopOrdersResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetStopOrdersResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.stopOrders.push(StopOrder.decode(reader, reader.uint32()));
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetStopOrdersResponse {
    return {
      stopOrders: globalThis.Array.isArray(object?.stopOrders)
        ? object.stopOrders.map((e: any) => StopOrder.fromJSON(e))
        : [],
    };
  },

  toJSON(message: GetStopOrdersResponse): unknown {
    const obj: any = {};
    if (message.stopOrders?.length) {
      obj.stopOrders = message.stopOrders.map((e) => StopOrder.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<GetStopOrdersResponse>): GetStopOrdersResponse {
    return GetStopOrdersResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetStopOrdersResponse>): GetStopOrdersResponse {
    const message = createBaseGetStopOrdersResponse();
    message.stopOrders = object.stopOrders?.map((e) => StopOrder.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;